+++ src/components/common/ui/layout/next-nav/NavBar.tsx
@@ -1,48 +1,37 @@
 'use client';
 
-import { useGetActiveReports } from '@/api/report';
 import { useGetUserAdmin, useUserQuery } from '@/api/user';
-import { hasActiveReport } from '@/hooks/useActiveReports';
+import useActiveLink from '@/hooks/useActiveLink';
+import useActiveReportNav from '@/hooks/useActiveReportNav';
 import { useControlScroll } from '@/hooks/useControlScroll';
+import useScrollDirection from '@/hooks/useScrollDirection';
 import { twMerge } from '@/lib/twMerge';
 import useAuthStore from '@/store/useAuthStore';
-import useScrollStore from '@/store/useScrollStore';
 import Link from 'next/link';
 import { usePathname, useRouter } from 'next/navigation';
-import { useEffect, useRef, useState } from 'react';
-
+import { useState } from 'react';
 import GlobalNavItem from '../header/GlobalNavItem';
 import GlobalNavTopBar from '../header/GlobalNavTopBar';
-import { SubNavItemProps } from '../header/SubNavItem';
 import KakaoChannel from '../nav/KakaoChannel';
 import SideNavItem from './SideNavItem';
-
-const scrollEventPage = [
-  '/report/landing',
-  '/program/challenge',
-  '/program/live',
-];
 
 const NavBar = () => {
   const router = useRouter();
   const pathname = usePathname() ?? '';
-  const lastScrollY = useRef(0);
 
   const [isOpen, setIsOpen] = useState(false);
-  const [reportItems, setReportItems] = useState<SubNavItemProps[]>([]);
-  const [activeLink, setActiveLink] = useState<
-    'HOME' | 'ABOUT' | 'PROGRAM' | 'ADMIN' | 'BLOG' | 'REPORT' | 'REVIEW' | ''
-  >('');
+
+  const activeLink = useActiveLink(pathname);
+  const reportNavList = useActiveReportNav();
+  const scrollDirection = useScrollDirection(pathname);
 
   const { isLoggedIn, logout } = useAuthStore();
-  const { setScrollDirection, scrollDirection } = useScrollStore();
 
   const { data: user } = useUserQuery({ enabled: isLoggedIn, retry: 1 });
   const { data: isAdmin } = useGetUserAdmin({
     enabled: isLoggedIn,
     retry: 1,
   });
-  const { data, isLoading } = useGetActiveReports();
 
   const toggleMenu = () => {
     setIsOpen(!isOpen);
@@ -51,103 +40,8 @@ const NavBar = () => {
     setIsOpen(false);
   };
 
-  useEffect(() => {
-    /* 활성화된 서류 진단만 서브 메뉴로 설정 */
-    const reportSubNavList: SubNavItemProps[] = [
-      {
-        children: '이력서 진단 받기',
-        href: '/report/landing/resume',
-        isNextRouter: true,
-      },
-      {
-        children: '자기소개서 진단 받기',
-        href: '/report/landing/personal-statement',
-        isNextRouter: true,
-      },
-      {
-        children: '포트폴리오 진단 받기',
-        href: '/report/landing/portfolio',
-        isNextRouter: true,
-      },
-      {
-        children: 'MY 진단서 보기',
-        href: '/report/management',
-        isNextRouter: true,
-        force: true,
-      },
-    ];
-
-    if (data) {
-      const navItems: SubNavItemProps[] = [];
-      const resumeInfoList = data?.resumeInfoList;
-      const personalStatementInfoList = data?.personalStatementInfoList;
-      const portfolioInfoList = data?.portfolioInfoList;
-
-      if (hasActiveReport(resumeInfoList)) {
-        navItems.push(reportSubNavList[0]);
-      }
-      if (hasActiveReport(personalStatementInfoList)) {
-        navItems.push(reportSubNavList[1]);
-      }
-      if (hasActiveReport(portfolioInfoList)) {
-        navItems.push(reportSubNavList[2]);
-      }
-      navItems.push(reportSubNavList[3]);
-      setReportItems(navItems);
-    } else {
-      setReportItems([reportSubNavList[3]]);
-    }
-  }, [data]);
-
   // 사이드바 열리면 스크롤 제한
   useControlScroll(isOpen);
-
-  useEffect(() => {
-    // Active 링크 설정
-    if (pathname.startsWith('/about')) {
-      setActiveLink('ABOUT');
-    } else if (pathname.startsWith('/program')) {
-      setActiveLink('PROGRAM');
-    } else if (pathname.startsWith('/admin')) {
-      setActiveLink('ADMIN');
-    } else if (pathname.startsWith('/blog')) {
-      setActiveLink('BLOG');
-    } else if (pathname.startsWith('/report')) {
-      setActiveLink('REPORT');
-    } else if (pathname.startsWith('/review')) {
-      setActiveLink('REVIEW');
-    } else if (location.pathname === '/') {
-      setActiveLink('HOME');
-    }
-  }, [pathname]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-
-    const handleScroll = () => {
-      // 현재 경로가 scrollEventPage 중 하나로 시작되지 않을 때는 스크롤 이벤트를 무시
-      if (!scrollEventPage.some((page) => location.pathname.startsWith(page))) {
-        return;
-      }
-
-      const currentScrollY = Math.max(0, window.scrollY);
-
-      if (currentScrollY > lastScrollY.current) {
-        setScrollDirection('DOWN');
-      } else if (currentScrollY < lastScrollY.current) {
-        setScrollDirection('UP');
-      }
-
-      lastScrollY.current = currentScrollY;
-    };
-
-    setScrollDirection('UP');
-    window.addEventListener('scroll', handleScroll);
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [pathname, setScrollDirection]);
 
   return (
     <>
@@ -179,9 +73,8 @@ const NavBar = () => {
               className="text-xsmall16"
               isNextRouter
               active={activeLink === 'REPORT'}
-              href={isLoading ? '#' : reportItems[0].href}
-              subNavList={reportItems}
-              subNavLoaded={!isLoading && !!data}
+              href={reportNavList.length === 0 ? '#' : reportNavList[0].href}
+              subNavList={reportNavList}
             >
               서류 피드백 REPORT
             </GlobalNavItem>
